Remove invalid option from esbuild's fromString
The `allowOverwrite` option only makes sense in `esbuild.build()`, and
throws an error when passed to `esbuild.transform`. To resolve this, use
different config object for the two calls. This issue went unnoticed due
to a bug in `fromHtmlMinifier()` where it didn't call `fromString` on
the provided CSS and JS minfiers at all.